requirements.txt
Django==2.0.6
django-ckeditor==5.6.1